+++ src/main/java/edu/sabanciuniv/main/Services/InstructerService.java
@@ -1,6 +1,7 @@
 package edu.sabanciuniv.main.Services;
 
 
+import edu.sabanciuniv.main.CustomException.GeneralCustomException;
 import edu.sabanciuniv.main.Entity.Course;
 import edu.sabanciuniv.main.Entity.Instructer;
 import edu.sabanciuniv.main.Repository.InstructerRepo;
@@ -28,15 +29,28 @@ public class InstructerService {
 
     }
 
-    public void deleteInstructer(int id) {
-        instructerRepo.deleteById(id);
+    public void deleteInstructer(int id) throws GeneralCustomException {
+        if (instructerRepo.findById(id).isPresent()) {
+            instructerRepo.deleteById(id);
+        } else {
+            throw new GeneralCustomException("Instructer not found");
+        }
+
     }
 
-    public void updateInstructer(Instructer instructer) {
-        instructerRepo.save(instructer);
+    public void updateInstructer(Instructer instructer) throws GeneralCustomException {
+        if (instructerRepo.findById(instructer.getId()).isPresent()) {
+            instructerRepo.save(instructer);
+        } else {
+            throw new GeneralCustomException("Instructer not found");
+        }
+
     }
 
-    public Instructer getInstructerById(int id) {
+    public Instructer getInstructerById(int id) throws GeneralCustomException {
+        if (!instructerRepo.findById(id).isPresent()) {
+            throw new GeneralCustomException("Instructer not found");
+        }
         return instructerRepo.findById(id).get();
     }
